Fixed a compiler error when building analytics unit tests for Android

+++ modules/juce_analytics/destinations/juce_ThreadedAnalyticsDestination.cpp
@@ -216,7 +216,7 @@ struct ThreadedAnalyticsDestinationTests   : public UnitTest
                              const std::deque<AnalyticsDestination::AnalyticsEvent>& b)
     {
         const auto numEntries = a.size();
-        expectEquals (b.size(), numEntries);
+        expectEquals ((int) b.size(), (int) numEntries);
 
         for (size_t i = 0; i < numEntries; ++i)
         {
